placed search, day, and shiftcard on the same row. Submitting PR
