+++ src/components/ManagerComponents/Shiftcard/Shiftcard.module.css
@@ -40,3 +40,7 @@
     float: left;
     margin-bottom: 10px;
 }
+
+.button {
+    float: right;
+}
