--- src/components/ManagerComponents/UnassignedShiftList/unassignedShiftList.module.css
@@ -1,32 +1,0 @@
-.assignShiftBox {
-    padding: 10%;
-}
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    padding-bottom: 3%;
-}
-
-.close{
-    text-align: right;
-}
-
-.assignShiftInfo{
-    display: flex;
-    justify-content: space-between;
-}
-
-.infoForShift{
-    background-color: #EEEEEE;
-    border-radius: 34px;
-    padding: 4px 10px 4px 10px;
-}
-
-.searchBar {
-  width: 53.6%;
-}
-
-.container {
-    flex-direction: row;
-  }
