+++ src/components/ManagerComponents/UnassignedShiftsList/UnassignedShiftsList.module.css
@@ -22,3 +22,12 @@
     border-radius: 34px;
     padding: 4px 10px 4px 10px;
 }
+.flex {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+}
+
+.flex2 {
+    display: flex;
+}
